first pass at cross-block state management
