+++ crates/flashblocks-rpc/Cargo.toml
@@ -72,6 +72,7 @@ metrics-derive.workspace = true
 eyre.workspace = true
 brotli.workspace = true
 arc-swap.workspace = true
+itertools.workspace = true
 
 [dev-dependencies]
 rand.workspace = true
