+++ crates/flashblocks-rpc/src/lib.rs
@@ -1,7 +1,7 @@
+pub mod fb_state;
 mod metrics;
-mod pending;
+mod pending_blocks;
 pub mod rpc;
-pub mod state;
 pub mod subscription;
 #[cfg(test)]
 mod tests;
